cart items count seting

+++ Frontent/src/components/Product/Product.jsx
@@ -7,15 +7,43 @@ import axios from 'axios'
 
 function Product(product) {
 
-
   const headers = useMemo(() => ({
     'Authorization': window.localStorage.getItem("token")
   }), [])
-  const deleteProduct = () => {
-    console.log(collections.server_base + product.removeLink)
-    axios.delete(collections.server_base + product.removeLink, { headers })
-    product.handleDeleteCallback(Date.now())
+
+  const deleteProduct = (e) => {
+    e.preventDefault()
+    if (window.confirm("are you sure to remove item from cart")) {
+      axios.delete(collections.server_base + product.removeLink, { headers })
+      product.handleDeleteCallback(Date.now())
+    }
   }
+
+  const handleIncrement = (e) => {
+    e.preventDefault()
+    axios.patch(`${collections.server_base}/add-to-cart/${product._id}/${1}/${product.userId}`).then((res) => {
+      if (res.data.cartPriceLimitErr) {
+        alert("Maximum 500,000 rupees in cart")
+      } else {
+        product.handleDeleteCallback(Date.now())
+      }
+    })
+  }
+
+  const handleDecrement = (e) => {
+    e.preventDefault()
+    if (product.count < 2) {
+      if (window.confirm("are you sure to remove item from cart")) {
+        axios.delete(collections.server_base + product.removeLink, { headers })
+        product.handleDeleteCallback(Date.now())
+      }
+    } else {
+      axios.patch(`${collections.server_base}/add-to-cart/${product._id}/${-1}/${product.userId}`).then((res) => {
+        product.handleDeleteCallback(Date.now())
+      })
+    }
+  }
+
   return (
     <div className='Product'>
       <div className='container-fluid'>
@@ -28,8 +56,8 @@ function Product(product) {
                   icon={<svg xmlns="http://www.w3.org/2000/svg" width="13" height="13" fill="currentColor" className="edit-bottun-svg bi bi-pencil" viewBox="0 0 16 16">
                     <path d="M12.146.146a.5.5 0 0 1 .708 0l3 3a.5.5 0 0 1 0 .708l-10 10a.5.5 0 0 1-.168.11l-5 2a.5.5 0 0 1-.65-.65l2-5a.5.5 0 0 1 .11-.168l10-10zM11.207 2.5 13.5 4.793 14.793 3.5 12.5 1.207 11.207 2.5zm1.586 3L10.5 3.207 4 9.707V10h.5a.5.5 0 0 1 .5.5v.5h.5a.5.5 0 0 1 .5.5v.5h.293l6.5-6.5zm-9.761 5.175-.106.106-1.528 3.821 3.821-1.528.106-.106A.5.5 0 0 1 5 12.5V12h-.5a.5.5 0 0 1-.5-.5V11h-.5a.5.5 0 0 1-.468-.325z" />
                   </svg>}
-                  title={product.cartItem ? "remove" :"delete"}
-                  width={product.cartItem ? "4.9rem" :"4.2rem"}
+                  title={product.cartItem ? "remove" : "delete"}
+                  width={product.cartItem ? "4.9rem" : "4.2rem"}
                   onClick={deleteProduct}
                 />
               </div>
@@ -47,6 +75,11 @@ function Product(product) {
                 {!product.deletedItem && <h5 className='product-price'>₹ {product?.price}/-<span className='product-mrp'>₹ {product.mrp}/-</span></h5>}
                 {!product.deletedItem && <h5>{product.category} <span className='product-offer' >{100 - (Math.round((parseInt(product?.price) * 100) / parseInt(product?.mrp)))}%  Offer</span> </h5>}
               </div>
+              {!isNaN(product.count) && <div className="product-count-div">
+                <div className="inc-div" onClick={handleIncrement} >+</div>
+                <div className="count-div">{product.count}</div>
+                <div className="dec-div" onClick={handleDecrement}>-</div>
+              </div>}
             </div>
           </div>
         </Link>
